+++ packages/expo-router/src/link/zoom/usePreventZoomTransitionDismissal.ios.tsx
@@ -4,10 +4,8 @@ import { use } from 'react';
 
 import { DescriptorsContext } from '../../fork/native-stack/descriptors-context';
 import { INTERNAL_EXPO_ROUTER_GESTURE_ENABLED_OPTION_NAME } from '../../navigationParams';
-import { NavigatorStateContext } from '../../react-navigation/core/useNavigationState';
 import { useRoute } from '../../react-navigation/native';
 import { useNavigation } from '../../useNavigation';
-import { isRoutePreloadedInStack } from '../../utils/stack';
 import { isModalPresentation } from '../../utils/stackPresentation';
 import { useSafeLayoutEffect } from '../../views/useSafeLayoutEffect';
 import { useIsPreview } from '../preview/PreviewRouteContext';
@@ -20,10 +18,9 @@ export function usePreventZoomTransitionDismissal(
   const context = use(ZoomTransitionTargetContext);
   const route = useRoute();
   const navigation = useNavigation();
-  const navigatorState = use(NavigatorStateContext);
   const isPreview = useIsPreview();
   const isFocused = navigation.isFocused();
-  const isPreloaded = isPreview ? false : isRoutePreloadedInStack(navigatorState, route);
+  const isPreloaded = !isPreview && route.isPreloaded === true;
 
   const descriptorsMap = use(DescriptorsContext);
   const currentDescriptor = descriptorsMap[route.key];